requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
librosa==0.10.1
numpy==1.24.3
pandas==2.0.3
pydub==0.25.1
SpeechRecognition==3.10.0
transformers==4.33.2
torch==2.0.1
scikit-learn==1.3.0
soundfile==0.12.1
python_speech_features
python_speech_features
spacy
plotly>=5.13.0
audio-recorder-streamlit>=0.0.8
streamlit-audio-recorder
matplotlib>=3.4.3
seaborn>=0.11.2
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
tqdm>=4.64.0
pytest>=7.0.0 